fix input_opt

+++ code/input_opt.py
@@ -62,7 +62,8 @@ class InputOptimizer(tf.keras.Model):
             print(out_numpy)
             # print(out_numpy.shape)
             # ax.imshow(out_numpy, vmin = -1, vmax = 1)
-            ax.imshow((out_numpy * 255).astype(np.uint8))
+            # ax.imshow((out_numpy * 255).astype(np.uint8))
+            ax.imshow(out_numpy*255)
         self.opt_imgs += [self.fig2img(fig)]
         plt.close(fig)
